Split utils into seperate classes for better project organisation
