+++ backend/reports/plate_recognition_utils.py
@@ -1,0 +1,33 @@
+import os
+import threading
+import time
+from datetime import datetime
+import requests
+
+_plate_recognizer_lock = threading.Lock()
+
+class PlateRecognitionError(Exception):
+    pass
+
+def recognize_plate(image_path):
+    api_key = os.environ.get("PLATE_RECOGNIZER_API_KEY")
+    if not api_key:
+        raise RuntimeError("❌ Plate Recognizer API key is not set in environment variables.")
+    with _plate_recognizer_lock:
+        time.sleep(1)
+        with open(image_path, 'rb') as img_file:
+            response = requests.post(
+                'https://api.platerecognizer.com/v1/plate-reader/',
+                files={'upload': img_file},
+                headers={'Authorization': f'Token {api_key}'}
+            )
+        if response.status_code not in (200, 201):
+            raise PlateRecognitionError(f"❌ API request failed: {response.status_code} - {response.text}")
+        data = response.json()
+        try:
+            plate = data['results'][0]['plate'].upper()
+            timestamp = datetime.now().strftime("%Y-%m-%d %H:%M:%S")
+            print(f"✅ {timestamp} — License plate recognized successfully: {plate}")
+            return plate
+        except (KeyError, IndexError):
+            raise PlateRecognitionError("❌ No license plate detected in the image.")
